requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pydicom>=2.4.0
matplotlib>=3.7.0
grad-cam>=1.4.8
streamlit>=1.30.0
